+++ Collision2D/Polygon.hpp
@@ -21,5 +21,5 @@ class Polygon : public VertexBasedBody {
     // Public methods
     sf::Vector2f getNormal(int i_index) override;
     pointSeparationData_type calculateMinPointSeparation(sf::Vector2f i_point) override;
-    int getNormalCount() override;
+    size_t getNormalCount() override;
 };
